// frontend/src/features/gateways/models/create-gateway-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gateways API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 *
 * @export
 * @interface CreateGatewayDto
 */
export interface CreateGatewayDto {
  /**
   *
   * @type {string}
   * @memberof CreateGatewayDto
   */
  serialNumber: string;
  /**
   *
   * @type {string}
   * @memberof CreateGatewayDto
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof CreateGatewayDto
   */
  ipAddress: string;
}